Fin problema 1_Examen2

+++ Problema1_Examen2/src/Vehiculo/Electrico.java
@@ -3,7 +3,7 @@ package Vehiculo;
 public class Electrico extends Vehiculo implements Impuesto {
     private double precioCarga;
 
-    public void setPrecioCarga(int precioCarga){
+    public void setPrecioCarga(double precioCarga){
         this.precioCarga = precioCarga;
     }
     public Electrico(String placa, String marca, String modelo, String color){
